Changes to get it working with VS2022

+++ VisualStudio/ProjectBase/ProjectNode.cs
@@ -2152,7 +2152,7 @@ namespace Microsoft.VisualStudio.Project
                 {
                     Array contextParamsAsArray = contextParams;
 
-                    int result = ivsExtensibility.RunWizardFile(wizardToRun, (int)dlgOwner, ref contextParamsAsArray, out wizResultAsInt);
+                    int result = ivsExtensibility.RunWizardFile(wizardToRun, dlgOwner, ref contextParamsAsArray, out wizResultAsInt);
 
                     if (!ErrorHandler.Succeeded(result) && result != VSConstants.OLE_E_PROMPTSAVECANCELLED)
                     {
